backend/requirements.txt
Babel==2.5.3
boto3==1.7.26
cairocffi==0.8.0
CairoSVG==2.1.3
cffi==1.11.5
coreapi==2.3.3
Django==2.0.5
django-anymail==2.2
django-compressor==2.2
django-cors-headers==2.2.0
django-crontab==0.7.1
django-libsass==0.7
django-rest-auth==0.9.3
django-storages==1.6.6
djangorestframework==3.8.2
dry-rest-permissions==0.1.10
# User github repo for Django Recurrence for Django 2.X support
git+https://github.com/django-recurrence/django-recurrence
glom==18.1.1
html2text==2018.1.9
Markdown==2.6.11
pendulum==2.0.1
psycopg2-binary==2.7.4
Pygments==2.2.0
pysftp==0.2.9
pytest-django==3.2.1
pytest==3.5.1
pytz==2017.3
requests==2.18.1
twilio==6.14.0
uWSGI==2.0.17
WeasyPrint==0.42.3
